refactor(dashboard-core): split New Session flow ownership
