+++ packages/dashboard-core/src/state/screen.ts
@@ -1,5 +1,5 @@
 import type { StationSnapshot } from "@station/contracts";
-import { reconcileNewSessionFlow } from "../flows/newSession.js";
+import { reconcileNewSessionFlow } from "../flows/newSession/reconciliation.js";
 import { selectMoveToGroupSessionContext } from "../selectors/selectors.js";
 import { reconcileDashboardFocus } from "./dashboardFocus.js";
 import { createEmptyTuiLocalRows, pruneLocalRowsForSnapshot } from "./localRows.js";
