+++ packages/dashboard-core/src/state/screens/fork.ts
@@ -6,7 +6,7 @@ import {
   editableTextInputIntentForInput,
   transitionEditableTextInput,
 } from "../../components/EditableTextInput/editing.js";
-import { createNewSessionNameToken } from "../../flows/newSession.js";
+import { createNewSessionNameToken } from "../../flows/newSession/names.js";
 import { selectDashboardSessionRow } from "../../selectors/dashboardSessionRows.js";
 import type { TuiKey } from "../keys.js";
 import { isReturnKey } from "../keys.js";
