--- packages/dashboard-core/test/unit/flows/newSession.test.ts
@@ -1,674 +1,0 @@
-import { describe, expect, it } from "vitest";
-import {
-  chooseNewSessionAgentById,
-  chooseNewSessionGroupById,
-  chooseNewSessionProjectById,
-  createNewSessionFlow,
-  createNewSessionNameToken,
-  newSessionActionEnabled,
-  newSessionIntentForAction,
-  newSessionIntentForInput,
-  reconcileNewSessionFlow,
-  transitionNewSessionFlow,
-  validateNewSessionCreate,
-} from "../../../src/flows/newSession.js";
-import { selectNewSessionHarnessOptions } from "../../../src/selectors/selectors.js";
-import {
-  createDashboardSnapshot,
-  createGroupedDashboardSnapshot,
-} from "../../fixtures/snapshots.js";
-
-describe("new session flow", () => {
-  it("defaults to the first configured project and first configured agent", () => {
-    const state = createNewSessionFlow(createHarnessSnapshot(), "k7p3x9");
-
-    expect(state).toEqual({
-      mode: "review",
-      reviewFocus: "create",
-      selectedProjectId: "web",
-      selectedHarness: "codex",
-      title: "web-k7p3x9",
-      branch: "web-k7p3x9",
-      titleSource: "generated",
-      groupSelection: { kind: "ungrouped" },
-      stepHistory: [],
-    });
-    expect(Object.hasOwn(state ?? {}, "draftName")).toBe(false);
-  });
-
-  it("creates deterministic path-safe name tokens from unique sources", () => {
-    expect(createNewSessionNameToken("source-a")).toMatch(/^[a-f0-9]{6}$/);
-    expect(createNewSessionNameToken("source-a")).toBe(createNewSessionNameToken("source-a"));
-    expect(createNewSessionNameToken("source-a")).not.toBe(createNewSessionNameToken("source-b"));
-  });
-
-  it("review menu: arrows cycle focus and enter activates the focused field", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    expect(opened.reviewFocus).toBe("create");
-
-    // Enter on the default "create" focus submits.
-    expect(newSessionIntentForInput(opened, input("\r", { return: true }))).toEqual({
-      type: "submit",
-    });
-
-    // Down cycles create → project (the top field); enter there opens the
-    // project picker, not submit.
-    expect(newSessionIntentForInput(opened, input("", { downArrow: true }))).toEqual({
-      type: "transition",
-      action: { type: "reviewFocus", dir: 1 },
-    });
-    const onProject = transitionNewSessionFlow(opened, { type: "reviewFocus", dir: 1 });
-    if (onProject?.mode !== "review") throw new Error("expected review");
-    expect(onProject.reviewFocus).toBe("project");
-    expect(newSessionIntentForInput(onProject, input("\r", { return: true }))).toEqual({
-      type: "transition",
-      action: { type: "pickProject" },
-    });
-
-    // Down again focuses Name; enter opens the name editor.
-    const onName = transitionNewSessionFlow(onProject, { type: "reviewFocus", dir: 1 });
-    if (onName?.mode !== "review") throw new Error("expected review");
-    expect(onName.reviewFocus).toBe("name");
-    expect(newSessionIntentForInput(onName, input("\r", { return: true }))).toEqual({
-      type: "transition",
-      action: { type: "editName" },
-    });
-
-    // Letter accelerators still work regardless of focus, including direct create.
-    expect(newSessionIntentForInput(onName, input("A"))).toEqual({
-      type: "transition",
-      action: { type: "pickAgent" },
-    });
-    expect(newSessionIntentForInput(onName, input("C"))).toEqual({ type: "submit" });
-  });
-
-  it("trims custom titles without changing the generated branch", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-
-    const editing = transitionNewSessionFlow(opened, { type: "editName" });
-    if (editing?.mode !== "editName") throw new Error("expected edit mode");
-
-    const state = typeName(editing, " Hexagonal PT 12! ");
-
-    expect(transitionNewSessionFlow(state, { type: "commitName" })).toMatchObject({
-      mode: "review",
-      title: "Hexagonal PT 12!",
-      branch: "web-aaaaaa",
-      titleSource: "custom",
-    });
-  });
-
-  it("keeps input interpretation out of the app input handler", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-
-    expect(newSessionIntentForInput(opened, input("P"))).toEqual({
-      type: "transition",
-      action: { type: "pickProject" },
-    });
-    expect(newSessionIntentForInput(opened, input("N"))).toEqual({
-      type: "transition",
-      action: { type: "editName" },
-    });
-    expect(newSessionIntentForInput(opened, input("G"))).toEqual({
-      type: "transition",
-      action: { type: "pickGroup" },
-    });
-    expect(newSessionIntentForInput(opened, input("E"))).toEqual({ type: "none" });
-    expect(newSessionIntentForInput(opened, input("p"))).toEqual({ type: "none" });
-    expect(newSessionIntentForInput(opened, input("a"))).toEqual({ type: "none" });
-    expect(newSessionIntentForInput(opened, input("\r", { return: true }))).toEqual({
-      type: "submit",
-    });
-
-    // In a pick step the shared selection engine owns ↑↓/↵/slot before this
-    // handler, so a slot key yields no intent here.
-    const picker = transitionNewSessionFlow(opened, { type: "pickAgent" });
-    if (picker?.mode !== "pickAgent") throw new Error("expected agent picker");
-    expect(newSessionIntentForInput(picker, input("2"))).toEqual({ type: "none" });
-    expect(chooseNewSessionAgentById(picker, snapshot, "opencode")).toMatchObject({
-      mode: "review",
-      selectedHarness: "opencode",
-    });
-  });
-
-  it("preselects only a same-project root Group", () => {
-    const snapshot = createGroupedDashboardSnapshot();
-
-    expect(
-      createNewSessionFlow(snapshot, "aaaaaa", { projectId: "web", groupId: "group_active" }),
-    ).toMatchObject({ groupSelection: { kind: "existing", groupId: "group_active" } });
-    expect(
-      createNewSessionFlow(snapshot, "aaaaaa", { projectId: "web", groupId: "group_build" }),
-    ).toMatchObject({ groupSelection: { kind: "ungrouped" } });
-    expect(
-      createNewSessionFlow(snapshot, "aaaaaa", { projectId: "web", groupId: "group_api" }),
-    ).toMatchObject({ groupSelection: { kind: "ungrouped" } });
-  });
-
-  it("commits an existing Group or a trimmed inline draft", () => {
-    const snapshot = createGroupedDashboardSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const picker = transitionNewSessionFlow(opened, { type: "pickGroup" });
-    if (picker?.mode !== "pickGroup") throw new Error("expected Group picker");
-
-    expect(chooseNewSessionGroupById(picker, snapshot, "group_active")).toMatchObject({
-      mode: "review",
-      reviewFocus: "group",
-      groupSelection: { kind: "existing", groupId: "group_active" },
-    });
-    expect(chooseNewSessionGroupById(picker, snapshot, "group_build")).toBe(picker);
-
-    const editing = transitionNewSessionFlow(picker, { type: "editGroupDraft" });
-    if (editing?.mode !== "editGroupDraft") throw new Error("expected Group editor");
-    expect(newSessionActionEnabled(snapshot, editing, "editGroupDraft.save")).toBe(false);
-    expect(newSessionActionEnabled(snapshot, editing, "editGroupDraft.back")).toBe(true);
-    const typed = "  Release  ".split("").reduce((state, value) => {
-      const next = applyInput(state, value);
-      if (next.mode !== "editGroupDraft") throw new Error("expected Group editor");
-      return next;
-    }, editing);
-    expect(newSessionActionEnabled(snapshot, typed, "editGroupDraft.save")).toBe(true);
-    expect(newSessionIntentForAction(typed, "editGroupDraft.save")).toEqual({
-      type: "transition",
-      action: { type: "commitGroupDraft" },
-    });
-    expect(newSessionIntentForAction(typed, "editGroupDraft.back")).toEqual({
-      type: "transition",
-      action: { type: "cancel" },
-    });
-    expect(applyInput(typed, "\r", { return: true })).toMatchObject({
-      mode: "review",
-      reviewFocus: "group",
-      groupSelection: { kind: "create", name: "Release" },
-    });
-  });
-
-  it("resets Group placement on project change and invalid snapshot replacement", () => {
-    const snapshot = createGroupedDashboardSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa", {
-      projectId: "web",
-      groupId: "group_active",
-    });
-    if (opened === undefined) throw new Error("expected a flow");
-    const projectPicker = transitionNewSessionFlow(opened, { type: "pickProject" });
-    if (projectPicker?.mode !== "pickProject") throw new Error("expected project picker");
-    expect(chooseNewSessionProjectById(projectPicker, snapshot, "api", "bbbbbb")).toMatchObject({
-      selectedProjectId: "api",
-      groupSelection: { kind: "ungrouped" },
-    });
-
-    const renamed = {
-      ...snapshot,
-      sessionGroups: snapshot.sessionGroups.map((group) =>
-        group.id === "group_active" ? { ...group, name: "Renamed", version: 2 } : group,
-      ),
-    };
-    expect(reconcileNewSessionFlow(opened, renamed).groupSelection).toEqual({
-      kind: "existing",
-      groupId: "group_active",
-    });
-
-    for (const invalid of [
-      {
-        ...snapshot,
-        sessionGroups: snapshot.sessionGroups.filter((group) => group.id !== "group_active"),
-      },
-      {
-        ...snapshot,
-        sessionGroups: snapshot.sessionGroups.map((group) =>
-          group.id === "group_active" ? { ...group, parentGroupId: "group_empty" } : group,
-        ),
-      },
-    ]) {
-      expect(reconcileNewSessionFlow(opened, invalid).groupSelection).toEqual({
-        kind: "ungrouped",
-      });
-    }
-  });
-
-  it("uses wizard history for substep cancellation", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-
-    const editing = transitionNewSessionFlow(opened, { type: "editName" });
-    expect(editing).toMatchObject({
-      mode: "editName",
-      stepHistory: ["review"],
-      draftName: { value: "", cursor: 0 },
-    });
-
-    const reviewed = transitionNewSessionFlow(editing ?? opened, { type: "cancel" });
-    expect(reviewed).toMatchObject({
-      mode: "review",
-      stepHistory: [],
-    });
-  });
-
-  it("keeps the chosen agent and regenerates generated names when the project changes", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-
-    // Pick a non-default harness so the assertion can tell "preserved" from "reset to default".
-    const chosen = { ...opened, selectedHarness: "opencode" as const };
-    const picker = transitionNewSessionFlow(chosen, { type: "pickProject" });
-    if (picker?.mode !== "pickProject") throw new Error("expected project picker");
-    const selected = chooseNewSessionProjectById(picker, snapshot, "api", "bbbbbb");
-
-    expect(selected).toMatchObject({
-      mode: "review",
-      selectedProjectId: "api",
-      selectedHarness: "opencode",
-      title: "api-bbbbbb",
-      branch: "api-bbbbbb",
-      titleSource: "generated",
-    });
-  });
-
-  it("keeps custom names when the project changes", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-
-    const custom = {
-      ...opened,
-      title: "Hexagonal PT 12",
-      titleSource: "custom" as const,
-    };
-    const picker = transitionNewSessionFlow(custom, { type: "pickProject" });
-    if (picker?.mode !== "pickProject") throw new Error("expected project picker");
-    const selected = chooseNewSessionProjectById(picker, snapshot, "api", "bbbbbb");
-
-    expect(selected).toMatchObject({
-      selectedProjectId: "api",
-      selectedHarness: "codex",
-      title: "Hexagonal PT 12",
-      branch: "api-bbbbbb",
-      titleSource: "custom",
-    });
-  });
-
-  it("ignores a project pick for an unknown id", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const picker = transitionNewSessionFlow(opened, { type: "pickProject" });
-    if (picker?.mode !== "pickProject") throw new Error("expected project picker");
-
-    expect(chooseNewSessionProjectById(picker, snapshot, "ghost", "bbbbbb")).toBe(picker);
-  });
-
-  it("commits a project by id from a larger list", () => {
-    const snapshot = createProjectSnapshot(10);
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const picker = transitionNewSessionFlow(opened, { type: "pickProject" });
-    if (picker?.mode !== "pickProject") throw new Error("expected project picker");
-
-    const selected = chooseNewSessionProjectById(picker, snapshot, "project-10", "bbbbbb");
-
-    expect(selected).toMatchObject({
-      mode: "review",
-      selectedProjectId: "project-10",
-      selectedHarness: "codex",
-      title: "project-10-bbbbbb",
-      branch: "project-10-bbbbbb",
-    });
-  });
-
-  it("yields no pick-step intent for any key (the selection engine owns them)", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const picker = transitionNewSessionFlow(opened, { type: "pickProject" });
-    if (picker?.mode !== "pickProject") throw new Error("expected project picker");
-
-    for (const key of [
-      input("0"),
-      input("2"),
-      input("j"),
-      input("", { downArrow: true }),
-      input("", { upArrow: true }),
-      input("\r", { return: true }),
-    ]) {
-      expect(newSessionIntentForInput(picker, key)).toEqual({ type: "none" });
-    }
-  });
-
-  it("ignores an agent pick for an unknown id", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const picker = transitionNewSessionFlow(opened, { type: "pickAgent" });
-    if (picker?.mode !== "pickAgent") throw new Error("expected agent picker");
-
-    expect(chooseNewSessionAgentById(picker, snapshot, "ghost")).toBe(picker);
-  });
-
-  it("uses vertical movement into the action row and horizontal movement within it", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const editing = transitionNewSessionFlow(opened, { type: "editName" });
-    if (editing?.mode !== "editName") throw new Error("expected edit mode");
-    expect(editing.editNameFocus).toBe("name");
-
-    const saveIntent = newSessionIntentForInput(editing, input("", { downArrow: true }));
-    expect(saveIntent).toEqual({
-      type: "transition",
-      action: { type: "editNameFocusSet", focus: "save" },
-    });
-    if (saveIntent.type !== "transition") throw new Error("expected transition");
-    const saveFocused = transitionNewSessionFlow(editing, saveIntent.action);
-    expect(saveFocused).toMatchObject({ mode: "editName", editNameFocus: "save" });
-    if (saveFocused?.mode !== "editName") throw new Error("expected edit mode");
-
-    expect(newSessionIntentForInput(saveFocused, input("x"))).toEqual({ type: "none" });
-    expect(newSessionIntentForInput(saveFocused, input("\r", { return: true }))).toEqual({
-      type: "transition",
-      action: { type: "commitName" },
-    });
-
-    const backIntent = newSessionIntentForInput(saveFocused, input("", { rightArrow: true }));
-    expect(backIntent).toEqual({
-      type: "transition",
-      action: { type: "editNameFocusSet", focus: "back" },
-    });
-    if (backIntent.type !== "transition") throw new Error("expected transition");
-    const backFocused = transitionNewSessionFlow(saveFocused, backIntent.action);
-    if (backFocused?.mode !== "editName") throw new Error("expected edit mode");
-    expect(backFocused.editNameFocus).toBe("back");
-    expect(newSessionIntentForInput(backFocused, input("\r", { return: true }))).toEqual({
-      type: "transition",
-      action: { type: "cancel" },
-    });
-    expect(newSessionIntentForInput(backFocused, input("", { upArrow: true }))).toEqual({
-      type: "transition",
-      action: { type: "editNameFocusSet", focus: "name" },
-    });
-    expect(newSessionIntentForInput(backFocused, input("", { escape: true }))).toEqual({
-      type: "transition",
-      action: { type: "cancel" },
-    });
-  });
-
-  it("resolves visible controls directly without generated input paths", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    expect(newSessionIntentForAction(opened, "review.project")).toEqual({
-      type: "transition",
-      action: { type: "pickProject" },
-    });
-    expect(newSessionIntentForAction(opened, "review.create")).toEqual({ type: "submit" });
-
-    const editing = transitionNewSessionFlow(opened, { type: "editName" });
-    if (editing?.mode !== "editName") throw new Error("expected edit mode");
-    const saveFocused = transitionNewSessionFlow(editing, {
-      type: "editNameFocusSet",
-      focus: "save",
-    });
-    if (saveFocused?.mode !== "editName") throw new Error("expected edit mode");
-    const focusName = newSessionIntentForAction(saveFocused, "editName.name");
-    expect(focusName).toEqual({
-      type: "transition",
-      action: { type: "editNameFocusSet", focus: "name" },
-    });
-    if (focusName.type !== "transition") throw new Error("expected transition");
-    expect(transitionNewSessionFlow(saveFocused, focusName.action)).toMatchObject({
-      mode: "editName",
-      editNameFocus: "name",
-    });
-    expect(newSessionIntentForAction(saveFocused, "editName.save")).toEqual({
-      type: "transition",
-      action: { type: "commitName" },
-    });
-    expect(newSessionIntentForAction(saveFocused, "editName.back")).toEqual({
-      type: "transition",
-      action: { type: "cancel" },
-    });
-  });
-
-  it("keeps Enter and Ctrl-S as predictable name-save paths", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const editing = transitionNewSessionFlow(opened, { type: "editName" });
-    if (editing?.mode !== "editName") throw new Error("expected edit mode");
-
-    expect(newSessionIntentForInput(editing, input("\r", { return: true }))).toEqual({
-      type: "transition",
-      action: { type: "commitName" },
-    });
-    expect(newSessionIntentForInput(editing, input("s", { ctrl: true }))).toEqual({
-      type: "transition",
-      action: { type: "commitName" },
-    });
-  });
-
-  it("moves the edit-name cursor and edits at the insertion point", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const editing = transitionNewSessionFlow(opened, { type: "editName" });
-    if (editing?.mode !== "editName") throw new Error("expected edit mode");
-
-    const typed = typeName(editing, "feature/foo");
-    const movedOnce = applyInput(typed, "", { leftArrow: true });
-    const movedTwice = applyInput(movedOnce, "", { leftArrow: true });
-    const movedLeft = applyInput(movedTwice, "", { leftArrow: true });
-    if (movedLeft?.mode !== "editName") throw new Error("expected edit mode");
-    expect(movedLeft.draftName.cursor).toBe(8);
-
-    const inserted = applyInput(movedLeft, "-bar");
-    expect(inserted).toMatchObject({
-      mode: "editName",
-      draftName: {
-        value: "feature/-barfoo",
-        cursor: 12,
-      },
-    });
-
-    const backspaced = applyInput(inserted, "", { backspace: true });
-    expect(backspaced).toMatchObject({
-      mode: "editName",
-      draftName: {
-        value: "feature/-bafoo",
-        cursor: 11,
-      },
-    });
-
-    const deleted = applyInput(backspaced, "", { delete: true });
-    expect(deleted).toMatchObject({
-      mode: "editName",
-      draftName: {
-        value: "feature/-baoo",
-        cursor: 11,
-      },
-    });
-  });
-
-  it("treats Ctrl-U as delete-before-cursor in edit-name", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const editing = transitionNewSessionFlow(opened, { type: "editName" });
-    if (editing?.mode !== "editName") throw new Error("expected edit mode");
-
-    const typed = typeName(editing, "featurefoo");
-    const movedOnce = applyInput(typed, "", { leftArrow: true });
-    const movedTwice = applyInput(movedOnce, "", { leftArrow: true });
-    const movedLeft = applyInput(movedTwice, "", { leftArrow: true });
-
-    expect(newSessionIntentForInput(movedLeft, input("u", { ctrl: true }))).toEqual({
-      type: "transition",
-      action: { type: "editNameInput", action: { type: "deleteBeforeCursor" } },
-    });
-    expect(applyInput(movedLeft, "u", { ctrl: true })).toMatchObject({
-      mode: "editName",
-      draftName: {
-        value: "foo",
-        cursor: 0,
-      },
-    });
-  });
-
-  it("maps left and right arrows to edit-name cursor movement", () => {
-    const snapshot = createHarnessSnapshot();
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-    const editing = transitionNewSessionFlow(opened, { type: "editName" });
-    if (editing?.mode !== "editName") throw new Error("expected edit mode");
-
-    expect(newSessionIntentForInput(editing, input("", { leftArrow: true }))).toEqual({
-      type: "transition",
-      action: { type: "editNameInput", action: { type: "moveCursor", delta: -1 } },
-    });
-    expect(newSessionIntentForInput(editing, input("", { rightArrow: true }))).toEqual({
-      type: "transition",
-      action: { type: "editNameInput", action: { type: "moveCursor", delta: 1 } },
-    });
-  });
-
-  it("orders agent options from configured harnesses without a project default", () => {
-    const snapshot = createHarnessSnapshot();
-    const api = snapshot.projects.find((project) => project.id === "api");
-    if (api === undefined) throw new Error("missing api project");
-
-    expect(selectNewSessionHarnessOptions(snapshot, api).map((option) => option.id)).toEqual([
-      "codex",
-      "opencode",
-      "scripted",
-    ]);
-  });
-
-  it("blocks unavailable agents while allowing degraded and unknown agents", () => {
-    const snapshot = createHarnessSnapshot({
-      codex: "unavailable",
-      opencode: "degraded",
-    });
-    const opened = createNewSessionFlow(snapshot, "aaaaaa");
-    if (opened === undefined) throw new Error("expected a flow");
-
-    expect(validateNewSessionCreate(snapshot, opened)).toMatchObject({
-      ok: false,
-      error: {
-        code: "HARNESS_PROVIDER_UNAVAILABLE",
-      },
-    });
-    expect(newSessionActionEnabled(snapshot, opened, "review.create")).toBe(false);
-
-    const opencode = { ...opened, selectedHarness: "opencode" };
-    expect(validateNewSessionCreate(snapshot, opencode)).toMatchObject({
-      ok: true,
-      harnessProvider: "opencode",
-    });
-    expect(newSessionActionEnabled(snapshot, opencode, "review.create")).toBe(true);
-
-    const unknown = { ...opened, selectedHarness: "scripted" };
-    expect(validateNewSessionCreate(snapshot, unknown)).toMatchObject({
-      ok: true,
-      harnessProvider: "scripted",
-    });
-  });
-});
-
-function createHarnessSnapshot(
-  statuses: Partial<
-    Record<"codex" | "opencode" | "scripted", "healthy" | "degraded" | "unavailable">
-  > = {},
-) {
-  const snapshot = createDashboardSnapshot();
-  return {
-    ...snapshot,
-    harnesses: [
-      { id: "codex", label: "codex" },
-      { id: "opencode", label: "opencode" },
-      { id: "scripted", label: "scripted" },
-    ],
-    providerHealth: {
-      ...snapshot.providerHealth,
-      codex: harnessHealth("codex", statuses.codex ?? "healthy", snapshot.generatedAt),
-      opencode: harnessHealth("opencode", statuses.opencode ?? "healthy", snapshot.generatedAt),
-    },
-  };
-}
-
-function createProjectSnapshot(count: number) {
-  const snapshot = createHarnessSnapshot();
-  const baseProject = snapshot.projects[0];
-  if (baseProject === undefined) throw new Error("expected project");
-  return {
-    ...snapshot,
-    projects: Array.from({ length: count }, (_, index) => {
-      const id = `project-${index + 1}`;
-      return {
-        ...baseProject,
-        id,
-        label: id,
-        root: `/tmp/station/${id}`,
-        defaults: {
-          ...baseProject.defaults,
-          harness: "codex",
-        },
-      };
-    }),
-    rows: [],
-    sessions: [],
-  };
-}
-
-function typeName(
-  initialState: NonNullable<ReturnType<typeof transitionNewSessionFlow>> & { mode: "editName" },
-  value: string,
-) {
-  return value.split("").reduce((state, input) => {
-    const next = applyInput(state, input);
-    if (next?.mode !== "editName") throw new Error("expected edit mode");
-    return next;
-  }, initialState);
-}
-
-function applyInput(
-  state: NonNullable<ReturnType<typeof transitionNewSessionFlow>>,
-  value: string,
-  key: Parameters<typeof newSessionIntentForInput>[1]["key"] = {},
-) {
-  const intent = newSessionIntentForInput(state, input(value, key));
-  if (intent.type !== "transition") throw new Error("expected transition intent");
-  const next = transitionNewSessionFlow(state, intent.action);
-  if (next === undefined) throw new Error("expected state");
-  return next;
-}
-
-function input(
-  value: string,
-  key: Parameters<typeof newSessionIntentForInput>[1]["key"] = {},
-): Parameters<typeof newSessionIntentForInput>[1] {
-  return {
-    input: value,
-    key,
-    token: "bbbbbb",
-  };
-}
-
-function harnessHealth(
-  providerId: string,
-  status: "healthy" | "degraded" | "unavailable",
-  lastCheckedAt: string,
-) {
-  return {
-    provider: providerId,
-    providerType: "harness" as const,
-    status,
-    lastCheckedAt,
-  };
-}
